converted getLocalCommands.js to ts
